Add support for multiple email sending methods
The `TestEmail` class in `TestEmail.cs` now supports sending emails using different methods based on the `Sender` environment variable. If the variable is not set or is set to "smtp", it uses the existing `SendAsync` method. If set to "azurecommunication", it uses a new `SendAzureCommunicationAsync` method. An `InvalidOperationException` is thrown for unsupported values.

A new class `AzureCommunicationSender` was added in `AzureCommunicationSender.cs` with a `SendAsync` method that currently throws a `NotImplementedException`.

An extension method `SendAzureCommunicationAsync` was added to the `CommonMailMessage` class in `AzureCommunicationSender.cs`, which creates an instance of `AzureCommunicationSender` and calls its `SendAsync` method.